// src/apis/models/Address.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Tag } from './Tag';
import { HttpFile } from '../http/http';

/**
 * Model for standardized address
 */
export class Address {
  /**
   * Address lines of the user's residence address
   */
  'addressLines': Array<any>;
  /**
   * Post code of the user's residence address
   */
  'postcode': string;
  /**
   * City of the user's residence address
   */
  'city': string;
  /**
   * State of the user's residence address
   */
  'state'?: string;
  /**
   * User's country of residence
   */
  'country': string;
  /**
   * Additional information that can be added via tags
   */
  'tags'?: Array<Tag>;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'addressLines',
      baseName: 'addressLines',
      type: 'Array<any>',
      format: '',
    },
    {
      name: 'postcode',
      baseName: 'postcode',
      type: 'string',
      format: '',
    },
    {
      name: 'city',
      baseName: 'city',
      type: 'string',
      format: '',
    },
    {
      name: 'state',
      baseName: 'state',
      type: 'string',
      format: '',
    },
    {
      name: 'country',
      baseName: 'country',
      type: 'string',
      format: '',
    },
    {
      name: 'tags',
      baseName: 'tags',
      type: 'Array<Tag>',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return Address.attributeTypeMap;
  }

  public constructor() {}
}