// src/apis/models/Assignment.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Assignment {
  'assigneeUserId': string;
  'assignedByUserId': string;
  'timestamp': number;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'assigneeUserId',
      baseName: 'assigneeUserId',
      type: 'string',
      format: '',
    },
    {
      name: 'assignedByUserId',
      baseName: 'assignedByUserId',
      type: 'string',
      format: '',
    },
    {
      name: 'timestamp',
      baseName: 'timestamp',
      type: 'number',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return Assignment.attributeTypeMap;
  }

  public constructor() {}
}